+++ cypress/apps/react-19-app/src/components/Datepicker.jsx
@@ -1,0 +1,7 @@
+
+export const Datepicker = () => {
+    return (
+        <sgds-datepicker dateformat="DD/MM/YYYY" mode="single">
+        </sgds-datepicker>
+    )
+}