+++ cypress/apps/react-19-app/src/components/IconButton.jsx
@@ -1,0 +1,6 @@
+
+export const IconButton = () => {
+    return (
+<sgds-icon-button name="plus"></sgds-icon-button>
+    )
+}