+++ cypress/apps/react-19-app/src/components/ProgressBar.jsx
@@ -1,0 +1,7 @@
+
+export const ProgressBar = () => {
+  return (
+    <sgds-progress-bar label="40%" value="40" ariamin="0" ariamax="100" arialabel="Loading in progress">
+  </sgds-progress-bar>
+  );
+};