+++ cypress/apps/react-19-app/src/components/Skeleton.jsx
@@ -1,0 +1,5 @@
+
+export const Skeleton = () => {
+    return    <sgds-skeleton width="200px" height="20px" borderradius="4px"></sgds-skeleton>
+
+}