+++ cypress/apps/react-19-app/src/components/Spinner.jsx
@@ -1,0 +1,5 @@
+
+export const Spinner = () => {
+    return <sgds-spinner>
+    </sgds-spinner>
+}