+++ cypress/apps/react-19-app/src/components/Textarea.jsx
@@ -1,0 +1,5 @@
+
+export const Textarea = () => {
+    return <sgds-textarea label="Label" name="textarea" maxlength="100">
+    </sgds-textarea>
+}